refactor : move components to common Banner, Slider, Uploader
