+++ src/components/common/Slider/SliderCard/index.tsx
