+++ src/components/common/Slider/index.tsx
