+++ src/components/common/Slider/slider.utils.ts
