+++ src/pages/create-haduri.tsx
@@ -1,4 +1,4 @@
-import HaduriFilterView from '@components/HaduriFilter';
+import HaduriFilter from '@components/HaduriFilter';
 import NextHead from '@components/common/NextHead';
 import { createHaduri } from '@components/common/NextHead/NextHead.utils';
 
@@ -6,7 +6,7 @@ const CreateHaduri = () => {
   return (
     <>
       <NextHead metaData={createHaduri} />
-      <HaduriFilterView />
+      <HaduriFilter />
     </>
   );
 };
